+++ app/pb_kits/playbook/pb_circle_icon_button/circle_icon_button.rb
@@ -14,8 +14,6 @@ module Playbook
       prop :disabled, type: Playbook::Props::Boolean,
                      default: false
       prop :icon, required: true
-      prop :dark, type: Playbook::Props::Boolean,
-                     default: false
       prop :link
       prop :new_window, type: Playbook::Props::Boolean,
                      default: false