+++ app/pb_kits/playbook/pb_distribution_bar/_distribution_bar.jsx
@@ -2,7 +2,7 @@
 
 import React from 'react'
 import classnames from 'classnames'
-import { spacing } from '../utilities/spacing.js'
+import { globalProps } from '../utilities/globalProps.js'
 
 type DistributionBarProps = {
   className?: String,
@@ -37,7 +37,7 @@ const DistributionBar = (props: DistributionBarProps) => {
   const normalizedValues = normalizeCharacters(widths)
 
   return (
-    <div className={classnames(`pb_distribution_bar_${size}`, spacing(props))}>
+    <div className={classnames(`pb_distribution_bar_${size}`, globalProps(props))}>
       {barValues(normalizedValues)}
     </div>
   )