+++ app/pb_kits/playbook/pb_table/_table.jsx
@@ -2,7 +2,7 @@
 
 import React, { type Node } from 'react'
 import classnames from 'classnames'
-import { spacing } from '../utilities/spacing.js'
+import { globalProps } from '../utilities/globalProps.js'
 
 type TableProps = {
   children: Array<Node> | Node,
@@ -43,7 +43,7 @@ const Table = (props: TableProps) => {
           'single-line': singleLine,
           'no-hover': disableHover,
         },
-        spacing(props)
+        globalProps(props)
       )}
     >
       {children}