+++ app/pb_kits/playbook/pb_button_toolbar/docs/_button_toolbar_secondary.jsx
@@ -1,0 +1,34 @@
+import React from 'react'
+import { Button, ButtonToolbar } from '../..'
+
+const ButtonToolbarSecondary = () => (
+  <div>
+    <ButtonToolbar
+        connected
+        orientation="horizontal"
+    >
+      <Button
+          text="Field"
+          variant="secondary"
+      />
+      <Button
+          text="Retail"
+          variant="secondary"
+      />
+      <Button
+          text="Event"
+          variant="secondary"
+      />
+      <Button
+          text="Training"
+          variant="secondary"
+      />
+      <Button
+          text="Not Working"
+          variant="secondary"
+      />
+    </ButtonToolbar>
+  </div>
+)
+
+export default ButtonToolbarSecondary